requirements.txt
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.6
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==24.2
django-nine==0.2.7
django-rest-elasticsearch==0.4.2
djangorestframework==3.15.1
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch==8.13.2
elasticsearch-dsl==8.13.1
inflection==0.5.1
packaging==24.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
